# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/authentication/invite":{"post":{"operationId":"AuthenticationController_inviteUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InviteUserDto"}}}},"responses":{"201":{"description":""}},"tags":["Authentication"]}},"/api/authentication/login":{"post":{"operationId":"AuthenticationController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["Authentication"]}},"/api/authentication/forgotPassword":{"post":{"operationId":"AuthenticationController_forgotPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordDto"}}}},"responses":{"200":{"description":""}},"tags":["Authentication"]}},"/api/authentication/recoverPassword":{"post":{"operationId":"AuthenticationController_recoverPassword","parameters":[{"name":"resetToken","required":true,"in":"header","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordDto"}}}},"responses":{"200":{"description":""}},"tags":["Authentication"]}},"/api/authentication/resetPassword":{"post":{"operationId":"AuthenticationController_resetPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}}},"responses":{"200":{"description":""}},"tags":["Authentication"]}},"/api/authentication/logout":{"post":{"operationId":"AuthenticationController_logout","parameters":[{"name":"Authorization","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Authentication"]}},"/api/property/details":{"post":{"operationId":"PropertyDetailsController_savePropertyDetails","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PropertyDetailsDto"}}}},"responses":{"201":{"description":""}},"tags":["Property Details"]}}},"info":{"title":"Fraxioned API","description":"Welcome to Fraxioned","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"UserPropertyDto":{"type":"object","properties":{"propertyID":{"type":"number"},"noOfShares":{"type":"string"},"acquisitionDate":{"format":"date-time","type":"string"}},"required":["propertyID","noOfShares","acquisitionDate"]},"InviteUserDto":{"type":"object","properties":{"email":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"addressLine1":{"type":"string"},"addressLine2":{"type":"string"},"state":{"type":"string"},"country":{"type":"string"},"city":{"type":"string"},"zipcode":{"type":"string"},"phoneNumber":{"type":"string"},"roleId":{"type":"number"},"updated_by":{"type":"number"},"created_by":{"type":"number"},"userPropertyDetails":{"$ref":"#/components/schemas/UserPropertyDto"}},"required":["email","firstName","lastName","addressLine1","addressLine2","state","country","city","zipcode","phoneNumber","roleId","updated_by","created_by","userPropertyDetails"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"ForgotPasswordDto":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]},"ChangePasswordDto":{"type":"object","properties":{"newPassword":{"type":"string"}},"required":["newPassword"]},"ResetPasswordDto":{"type":"object","properties":{"oldPassword":{"type":"string"},"newPassword":{"type":"string"},"userId":{"type":"number"}},"required":["oldPassword","newPassword","userId"]},"PropertyDetailsDto":{"type":"object","properties":{"no_of_guests_allowed":{"type":"number"},"no_of_bedrooms":{"type":"number"},"no_of_bathrooms":{"type":"number"},"no_of_pets_allowed":{"type":"number"},"square_footage":{"type":"string"},"check_in_time":{"format":"date-time","type":"string"},"check_out_time":{"type":"number"},"pet_policy":{"type":"string"},"fee_per_pet":{"type":"number"},"cleaning_fee":{"type":"number"},"peakSeasonStartDate":{"type":"number"},"peakSeasonEndDate":{"type":"number"},"peakSeasonAllottedNights":{"type":"number"},"offSeasonAllottedNights":{"type":"number"},"peakSeasonAllottedHolidayNights":{"type":"number"},"offSeasonAllottedHolidayNights":{"type":"number"},"lastMinuteBookingAllottedNights":{"type":"number"},"wifi_network":{"type":"string"},"created_by":{"format":"date-time","type":"string"},"updated_by":{"format":"date-time","type":"string"},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"}},"required":["no_of_guests_allowed","no_of_bedrooms","no_of_bathrooms","no_of_pets_allowed","square_footage","check_in_time","check_out_time","pet_policy","fee_per_pet","cleaning_fee","peakSeasonStartDate","peakSeasonEndDate","peakSeasonAllottedNights","offSeasonAllottedNights","peakSeasonAllottedHolidayNights","offSeasonAllottedHolidayNights","lastMinuteBookingAllottedNights","wifi_network","created_by","updated_by","created_at","updated_at"]}}}}